# swagger.json
{
    
    "openapi": "3.0.0",
    "info": {
      "version": "1.0.0",
      "title": "Library API",
      "description": "API for managing books and members in a library"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": [
      "http"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],

    "paths": {

      "/books": {
        "get": {
          "tags": [
            "check the book"
          ],
          "summary": "Get all books available",
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "array",
                "example": 
                {
                    "code": "JK-45",
                    "title": "Harry Potter",
                    "author": "J.K Rowling",
                    "stock": 1
                }
              }
            }
          }
        }
        
    },"/members" :{
        "get": {
            "tags": [
              "member check"
            ],
            "summary": "Get all members",
            "responses": {
              "200": {
                "description": "Successful response",
                "schema": {
                  "type": "array",
                  "example": 
                  [
                    {
                        "code": "M001",
                        "name": "Angga"
                    },
                    {
                        "code": "M002",
                        "name": "Ferry"
                    },
                    {
                        "code": "M003",
                        "name": "Putri"
                    }
                ]
                }
              }
            }
          }

    }, 
"/borrow-book" : { 
  "post": {
    "tags": [
      "borrow-book"
    ],
    "summary": "Borrow book by input member_code and book_code",
    "description": "Add member_code and book_code",
    "operationId": "borrowBook",
    "requestBody":  {
      "description": "Create borrowing transaction",
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "book_code": {
                "type": "string"},
                "member_code": {
                  "type": "string"
                }
              
            }
          }
        }
      },
      "required": true
    } ,
    "responses": {
      "200": {
        "description": "Successful operation",
        "content": {
          "application/json": {
            "schema": {
             "type": "code 200"
            }
          }
        }
      },
      "405": {
        "description": "Invalid input"
      }
    }
  }
}, 
"/return-book" : { 
  "post": {
    "tags": [
      "return-book"
    ],
    "summary": "Return book by input member_code and book_code",
    "description": "Add member_code and book_code",
   
    "requestBody": {
      "description": "Create returning transaction",
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "book_code": {
                "type": "string"},
                "member_code": {
                  "type": "string"
                }
              
            }
          }
        }
      },
      "required": true
    },
    "responses": {
      "200": {
        "description": "Successful operation",
        "content": {
          "application/json": {
            "schema": {
             "type": "code 200"
            }
          }
        }
      },
      "405": {
        "description": "Invalid input"
      }
    }
  }
},
"/update-penalized" : { 
  "post": {
    "tags": [
      "Update penalized state manually"
    ],
    "summary": "Update penalized state",
    "description": "Penalized can be automatic but this method is for update penalized state manually for debugging",
   
    "requestBody": {
      "description": "Update penalized",
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
                "member_code": {
                  "type": "string"
                }
              
            }
          }
        }
      },
      "required": true
    },
    "responses": {
      "200": {
        "description": "Successful operation",
        "content": {
          "application/json": {
            "schema": {
             "type": "code 200"
            }
          }
        }
      },
      "405": {
        "description": "Invalid input"
      }
    }
  }
}}}

    
    

    